Make sure reference bom-refs are consistent
Commit 6b4e52f49021f72c4bb8d6 changed the references bom-refs to
use a PURL of the requirement.  This basically broke the SBOM
because the components they referred to where using the PURL of
the actual version seen as their bom-ref.

This changes the bom-refs of reference components to their
requirement: the actual version of the component should be
intospected from their "version" value.